refactor(widget): remove duplicate actions/sagas/reducer/selector of bizlogic
