+++ webapp/app/containers/Widget/reducer.ts
@@ -35,9 +35,6 @@ import {
   ADD_WIDGET_FAILURE
 } from './constants'
 import {
-  LOAD_BIZLOGICS,
-  LOAD_BIZLOGICS_SUCCESS,
-  LOAD_BIZLOGICS_FAILURE,
   LOAD_BIZDATAS,
   LOAD_BIZDATAS_SUCCESS,
   LOAD_BIZDATAS_FAILURE,
@@ -48,8 +45,7 @@ import { fromJS } from 'immutable'
 const initialState = fromJS({
   widgets: false,
   bizdatasLoading: false,
-  bizdatas: false,
-  bizlogics: false
+  bizdatas: false
 })
 
 function widgetReducer (state = initialState, action) {
@@ -108,10 +104,6 @@ function widgetReducer (state = initialState, action) {
       return state.set('bizdatasLoading', false)
     case CLEAR_BIZDATAS:
       return state.set('bizdatas', false)
-    case LOAD_BIZLOGICS_SUCCESS:
-      return state.set('bizlogics', payload.bizlogics)
-    case LOAD_BIZLOGICS_FAILURE:
-      return state
     default:
       return state
   }
