+++ webapp/app/containers/Widget/selectors.ts
@@ -27,11 +27,6 @@ const makeSelectWidgets = () => createSelector(
   (widgetState) => widgetState.get('widgets')
 )
 
-const makeSelectBizlogics = () => createSelector(
-  selectWidget,
-  (widgetState) => widgetState.get('bizlogics')
-)
-
 const makeSelectBizdatas = () => createSelector(
   selectWidget,
   (widgetState) => widgetState.get('bizdatas')
@@ -45,7 +40,6 @@ const makeSelectBizdatasLoading = () => createSelector(
 export {
   selectWidget,
   makeSelectWidgets,
-  makeSelectBizlogics,
   makeSelectBizdatas,
   makeSelectBizdatasLoading
 }
